Photos information directly from mediastore for real time update
