+++ app/build.gradle
@@ -52,4 +52,10 @@ dependencies {
     compile 'com.mikepenz:iconics-core:2.5.5@aar'
     compile 'com.mikepenz:google-material-typeface:2.1.0.1.original@aar'
     compile 'com.mikepenz:fontawesome-typeface:4.5.0.1@aar'
+    /* compile('com.github.afollestad.material-dialogs:core:0.8.5.3@aar') {
+         transitive = true
+     }
+     compile('com.github.afollestad.material-dialogs:commons:0.8.5.3@aar') {
+         transitive = true
+     }*/
 }
