+++ app/src/main/java/com/leafpic/app/Adapters/PhotosAdapter.java
@@ -10,7 +10,6 @@ import android.widget.TextView;
 import com.bumptech.glide.Glide;
 import com.leafpic.app.Photo;
 import com.leafpic.app.R;
-import com.leafpic.app.utils.string;
 
 import java.util.ArrayList;
 
@@ -40,18 +39,27 @@ public class PhotosAdapter extends RecyclerView.Adapter<PhotosAdapter.ViewHolder
     @Override
     public void onBindViewHolder(PhotosAdapter.ViewHolder holder, int position) {
         Photo f = photos.get(position);
-        Glide.with(holder.imageView.getContext())
-                .load(f.Path)
-                .asBitmap()/* TODO gif previews */
-                .centerCrop()
-                .placeholder(R.drawable.ic_empty)
-                //.crossFade()
-                .into(holder.imageView);
 
-        String mime = string.getMimeType(f.Path);/* TODO mime type form media storage */
-        if (mime != null && mime.equals("image/gif"))
+
+        if (f.MIME.equals("image/gif")) {
+            Glide.with(holder.imageView.getContext())
+                    .load(f.Path)
+                    .asGif()
+                    .centerCrop()
+                    .placeholder(R.drawable.ic_empty)
+                    //.crossFade()
+                    .into(holder.imageView);
             holder.gifIcon.setVisibility(View.VISIBLE);
-        else holder.gifIcon.setVisibility(View.INVISIBLE);
+        } else {
+            holder.gifIcon.setVisibility(View.INVISIBLE);
+            Glide.with(holder.imageView.getContext())
+                    .load(f.Path)
+                    .asBitmap()
+                    .centerCrop()
+                    .placeholder(R.drawable.ic_empty)
+                    //.crossFade()
+                    .into(holder.imageView);
+        }
 
         holder.path.setTag(f.Path);
         if (f.isSelected()) {
