+++ app/src/main/java/com/leafpic/app/Photo.java
@@ -1,13 +1,9 @@
 package com.leafpic.app;
 
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.os.Parcel;
 import android.os.Parcelable;
-import com.leafpic.app.utils.string;
 
 public class Photo implements Parcelable {
-
 
     @SuppressWarnings("unused")
     public static final Parcelable.Creator<Photo> CREATOR = new Parcelable.Creator<Photo>() {
@@ -22,64 +18,44 @@ public class Photo implements Parcelable {
         }
     };
     public String Path;
-    String ID;
-    String FolderPath;
+    public String MIME;
     String DateTaken;
+    String FolderPath;
     String name;
-    boolean selected;
+    boolean selected = false;
 
     public Photo(String path) {
         Path = path;
-        FolderPath = string.getBucketPathbyImagePath(path);
-        name = string.getPhotoNamebyPath(path);
     }
 
     public Photo(String path, String dateTaken) {
         Path = path;
-        FolderPath = string.getBucketPathbyImagePath(path);
-        name = string.getPhotoNamebyPath(path);
         DateTaken = dateTaken;
     }
 
 
-    public Photo(String nome, String folderPath, String dateTaken) {
+    //hidden util
+    /*public Photo(String nome, String folderPath, String dateTaken) {
 
         Path = string.getPhotoPathByFolderPathAndName(folderPath, nome);
         FolderPath = folderPath;
         DateTaken = dateTaken;
         name = nome;
+    }*/
+
+    public Photo(String path, String dateTaken, String mime) {
+        Path = path;
+        DateTaken = dateTaken;
+        MIME = mime;
     }
 
     protected Photo(Parcel in) {
         Path = in.readString();
-        ID = in.readString();
-        FolderPath = in.readString();
         DateTaken = in.readString();
+        MIME = in.readString();
+        FolderPath = in.readString();
         name = in.readString();
         selected = in.readByte() != 0x00;
-    }
-
-    public static int calculateInSampleSize(
-            BitmapFactory.Options options, int reqWidth, int reqHeight) {
-        // Raw height and width of image
-        final int height = options.outHeight;
-        final int width = options.outWidth;
-        int inSampleSize = 1;
-
-        if (height > reqHeight || width > reqWidth) {
-
-            final int halfHeight = height / 2;
-            final int halfWidth = width / 2;
-
-            // Calculate the largest inSampleSize value that is a power of 2 and keeps both
-            // height and width larger than the requested height and width.
-            while ((halfHeight / inSampleSize) > reqHeight
-                    && (halfWidth / inSampleSize) > reqWidth) {
-                inSampleSize *= 2;
-            }
-        }
-
-        return inSampleSize;
     }
 
     public boolean isSelected() {
@@ -90,16 +66,6 @@ public class Photo implements Parcelable {
         selected = val;
     }
 
-    public int getDominantColor() {
-        final BitmapFactory.Options options = new BitmapFactory.Options();
-        options.inJustDecodeBounds = true;
-        BitmapFactory.decodeFile(Path, options);//.decodeResource(res, resId, options);
-        options.inSampleSize = calculateInSampleSize(options, 1, 1);
-        options.inJustDecodeBounds = false;
-        Bitmap b = BitmapFactory.decodeFile(Path, options);
-        return b.getPixel(0, 0);
-    }
-
     @Override
     public int describeContents() {
         return 0;
@@ -108,9 +74,9 @@ public class Photo implements Parcelable {
     @Override
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeString(Path);
-        dest.writeString(ID);
-        dest.writeString(FolderPath);
         dest.writeString(DateTaken);
+        dest.writeString(MIME);
+        dest.writeString(FolderPath);
         dest.writeString(name);
         dest.writeByte((byte) (selected ? 0x01 : 0x00));
     }
